chore: cleanup dead code
